Move scancodes to their own file.
