+++ src/scancodes/mod.rs
@@ -1,0 +1,7 @@
+//! A collection of Scancode implementations
+
+mod set1;
+mod set2;
+
+pub use self::set1::ScancodeSet1;
+pub use self::set2::ScancodeSet2;
